fix: возвращает старую реализацию сервисов, добавляет пример для Studio
