+++ src/domain/exceptions/base.py
@@ -1,0 +1,2 @@
+class BBBException(Exception):
+    ...
