+++ src/domain/exceptions/studio.py
@@ -1,0 +1,7 @@
+from sqlalchemy.exc import NoResultFound
+
+from src.domain.exceptions.base import BBBException
+
+
+class StudioNotFoundException(BBBException, NoResultFound):
+    ...
