--- src/domain/models/repositories/orm/operations/__init__.py
@@ -1,4 +1,0 @@
-from .create import ORMCreate
-from .delete import ORMDelete
-from .read import ORMReadOne, ORMReadAll
-from .update import ORMUpdate
