+++ src/domain/models/repositories/studio.py
@@ -1,9 +1,8 @@
-from src.domain.models.repositories.orm.repository import ORMRepository
+from src.domain.models import Studio
+from src.domain.models.repositories.SQLAlchemy import SQLAlchemyRepository
 from src.domain.schemas.studio import StudioCreate, StudioUpdate
-from src.models import Studio
 
 
-class StudioRepository(ORMRepository
-                       [Studio, StudioCreate, StudioUpdate]):
+class StudioRepository(SQLAlchemyRepository[Studio, StudioCreate, StudioUpdate]):
     def __init__(self):
         super().__init__(Studio)
