+++ src/domain/schemas/studio.py
@@ -1,0 +1,17 @@
+from src.domain.schemas.base import BaseSchema
+
+
+class StudioBase(BaseSchema):
+    ...
+
+
+class StudioRead(StudioBase):
+    ...
+
+
+class StudioCreate(StudioBase):
+    ...
+
+
+class StudioUpdate(StudioBase):
+    ...
