+++ src/domain/services/studio.py
@@ -1,14 +1,10 @@
-from src.domain.models.repositories.studio import StudioRepository
-from src.domain.services.usecases.studio import DeleteStudioByIdUseCase, CreateStudioUseCase, \
-    UpdateStudioByIdUseCase, GetStudioByIdUseCase, GetStudioByNameUseCase, GetAllStudiosUseCase
+from src.domain.exceptions.studio import StudioNotFoundException
 from src.domain.models import Studio
+from src.domain.models.repositories.studio import StudioRepository
+from src.domain.schemas.studio import StudioCreate, StudioUpdate
+from src.domain.services.base import ModelService
 
 
-class StudioService(CreateStudioUseCase,
-                    GetStudioByIdUseCase,
-                    GetStudioByNameUseCase,
-                    GetAllStudiosUseCase,
-                    UpdateStudioByIdUseCase,
-                    DeleteStudioByIdUseCase):
+class StudioService(ModelService[StudioRepository, Studio, StudioCreate, StudioUpdate]):
     def __init__(self):
-        super().__init__(Studio, StudioRepository)
+        super().__init__(StudioRepository(), StudioNotFoundException)
