--- src/domain/services/usecases/common/get_all.py
@@ -1,9 +1,0 @@
-from src.domain.models.base import BaseModel
-from src.domain.models.repositories.orm.repository import ORMRepository
-from src.infrastructure.item.usecases.read import GetAllItemsUseCase
-
-
-class GetAllUseCase[Model: BaseModel,
-                    Repository: ORMRepository] \
-            (GetAllItemsUseCase[Model, Repository]):
-    ...
