--- src/infrastructure/item/operations/__init__.py
@@ -1,4 +1,0 @@
-from .create import CreateOperation
-from .read import GetOneOperation, GetAllOperation
-from .update import UpdateOperation
-from .delete import DeleteOperation
