--- src/infrastructure/item/operations/base.py
@@ -1,8 +1,0 @@
-from abc import ABC
-
-
-class RepositoryOperation[Item](ABC):
-    _model: type[Item]
-
-    def __init__(self, model: type[Item]):
-        self._model = model
