--- src/infrastructure/item/operations/update.py
@@ -1,11 +1,0 @@
-from abc import ABC, abstractmethod
-
-from src.infrastructure.item.operations.base import RepositoryOperation
-
-
-class UpdateOperation[Item, UpdateSchema](RepositoryOperation, ABC):
-    @abstractmethod
-    async def update(self,
-                     schema: UpdateSchema,
-                     *where) -> list[Item]:
-        ...
