+++ src/main.py
@@ -11,7 +11,7 @@ def include_routers(app_: FastAPI, *routers: APIRouter) -> None:
 @asynccontextmanager
 async def lifespan(fastapi_app: FastAPI):
     """Запускаем код до и после запуска приложения"""
-    include_routers(fastapi_app, auth_router, user_router)
+    include_routers(fastapi_app, auth_router, studio_router, user_router)
     yield  # Возвращаем работу приложению
     # Тут можно выполнить код после завершения приложения
